增加一些tooltip；修复Mutable机器只搜索单一配方的问题;修复纳米核心结构错误;部分Mutable机器使用Add的均分跨配方逻辑;Record与部分逻辑位置移动

+++ src/main/java/com/gtladd/gtladditions/api/machine/logic/MutableRecipesLogic.java
@@ -172,7 +172,7 @@ public class MutableRecipesLogic<T extends WorkableElectricMultiblockMachine & I
             match = IParallelLogic.getRecipeOutputChance(machine, match);
             if (RecipeRunnerHelper.handleRecipeInput(machine, match)) {
                 remain -= pair.secondLong();
-                totalEu += (long) (getRecipeEut(match) * match.duration * euMultiplier);
+                totalEu += getRecipeEut(match) * match.duration * euMultiplier;
                 var item = match.outputs.get(ItemRecipeCapability.CAP);
                 if (item != null) itemOutputs.addAll(item);
                 var fluid = match.outputs.get(FluidRecipeCapability.CAP);
